pyproject.toml
[tool.poetry]
name = "tradingview-api"
version = "0.1.0"
description = "Python port of TradingView API wrapper"
readme = "README.md"
authors = ["minh <minh@local>"]
packages = [{ include = "tradingview_api" }]

[tool.poetry.dependencies]
python = ">=3.9"
requests = "^2.31.0"
websocket-client = "^1.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
